feat(popover): add Wrapper usage to Popover
