+++ packages/tailor-ui/src/Tooltip/Tooltip.tsx
@@ -1,4 +1,4 @@
-import React, { FC, ReactNode, useRef } from 'react';
+import React, { ComponentType, FC, ReactNode, useRef } from 'react';
 
 import {
   useHoverTrigger,
@@ -13,7 +13,7 @@ import TooltipPopup from './TooltipPopup';
 import { StyledTooltipProps } from './styles';
 
 export type TooltipProps = StyledTooltipProps & {
-  Wrapper?: FC;
+  Wrapper?: ComponentType;
   /**
    * Whether the floating tooltip card is visible by default. Only support when the trigger is `click`
    */
